moving StateContainer into its own folder and css file
