+++ src/containers/StateContainer/StateContainer.jsx
@@ -44,7 +44,7 @@ const StateContainer = ({ stateDiffs, clearLog }) => {
             </span>
           </div>
         );
-      })}
+      }).reverse()}
     </div>
   );
 };
